+++ backend/src/features/accounts/account.controller.ts
@@ -9,7 +9,9 @@ import {
 
 import type { AccountRow } from "./../../core/database/types/account-row.type.ts";
 import { RiotApiService } from "./../../integrations/riot/riot-api.service.ts";
+import { splitMatchId } from "./../../lib/split-match-id.ts";
 import { AccountService } from "./account.service.ts";
+import { AccountMatchService } from "./account-matches/account-match.service.ts";
 
 /** The response shape for an account, in the API's own casing. */
 const toResponse = (account: AccountRow) => ({
@@ -22,10 +24,16 @@ const toResponse = (account: AccountRow) => ({
 export class AccountController {
 	private readonly logger = new Logger(AccountController.name);
 	private readonly accounts: AccountService;
+	private readonly accountMatches: AccountMatchService;
 	private readonly riot: RiotApiService;
 
-	constructor(accounts: AccountService, riot: RiotApiService) {
+	constructor(
+		accounts: AccountService,
+		accountMatches: AccountMatchService,
+		riot: RiotApiService,
+	) {
 		this.accounts = accounts;
+		this.accountMatches = accountMatches;
 		this.riot = riot;
 	}
 
@@ -67,4 +75,51 @@ export class AccountController {
 		this.logger.debug(`Fetched ${entries.length} ranked entries`);
 		return entries;
 	}
+
+	/**
+	 * `GET /accounts/:puuid/matches?before=&start=&count=` — a page of the
+	 * account's match ids, newest first.
+	 *
+	 * `before` is a match id and the cursor form; it is what a client paging
+	 * through a history should send, because `start` counts from a head that
+	 * moves as games finish. `start` still works, and the response is a plain
+	 * array of ids either way.
+	 */
+	@Get(":puuid/matches")
+	async getMatchIds(
+		@Param("puuid") puuid: string,
+		@Query("start") start?: string,
+		@Query("count") count?: string,
+		@Query("before") before?: string,
+	) {
+		const startIndex = Number(start) || 0;
+		const pageSize = Number(count) || 5;
+
+		const matchIds = await this.accountMatches.getMatchIds(puuid, {
+			beforeGameId: parseCursor(before),
+			start: startIndex,
+			count: pageSize,
+		});
+		this.logger.debug(
+			`Served ${matchIds.length} match ids (before=${before ?? "-"}, start=${startIndex}, count=${pageSize})`,
+		);
+		return matchIds;
+	}
 }
+
+/**
+ * Reads the `before` cursor as the game id to page from.
+ *
+ * The cursor is a match id rather than a bare game id so a client can hand back
+ * the last id it was served without taking it apart. It is only ever split
+ * here — the id itself is never looked up, so a cursor naming a match the index
+ * does not hold still pages from the right position.
+ */
+const parseCursor = (before: string | undefined) => {
+	if (!before) return undefined;
+	try {
+		return splitMatchId(before).gameId;
+	} catch {
+		throw new BadRequestException(`Not a match id: ${before}`);
+	}
+};
